% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mostdepth_tn.R
\name{plot_mosdepth_tn}
\alias{plot_mosdepth_tn}
\title{Plot results from mosdepth output for Tumor/Normal pairs}
\usage{
plot_mosdepth_tn(t_bed = NULL, n_bed = NULL, plot_file = NULL, col = NULL)
}
\arguments{
\item{t_bed}{mosdepth output from tumor}

\item{n_bed}{mosdepth output from normal}

\item{plot_file}{basename for output plot}

\item{col}{Colors. Default c("#95a5a6", "#7f8c8d")}
}
\description{
Plot results from mosdepth output for Tumor/Normal pairs
}
